New inheritance scheme for connections.
This modifies the mixin pattern for composing the `connection_base`,
`basic_connection`, the `connectionpolicy` classes, and the concrete
connection classes.  Instead of _containing_ an object of one of the
policy classes, a concrete connection class is now _derived from_ its
policy class.  (This is not normally a good thing, but it happens to
simplify the setup and avoids nasty pitfalls in construction order.)

The interface for defining connection types is modified, so this is
in principle a source-incompatible change.  In practice, I doubt
anyone actually defines their own connection types, in which case it
won't matter.

With this change, you can now delete a connection object without
exact static knowledge of its type.